sort and Dividing the search into two components
