+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 // import { Dropdown } from 'react-bootstrap';
 import './App.css';
-import {Guides,Home,LogIn,SingUp} from './components/pages'
+import { Guides, Home, LogIn, SingUp } from './components/pages'
 import Navbar from './components/Navbar'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { library } from '@fortawesome/fontawesome-svg-core'
