+++ client/src/components/heroSection.jsx
@@ -3,7 +3,7 @@ import '../App.css';
 import { Button } from './Button';
 import './css/heroSection.css';
 import Search from './search'
-
+// import List from './LIstGuides';
 
 function HeroSection({history}) {
   return (
