+++ client/src/components/pages/index.js
@@ -1,5 +1,5 @@
 
-import Guides from './listGuides'
+import Guides from './Guides'
 import Home from './Home'
 import LogIn from './logIn'
 import SingUp from './singUp'
