--- client/src/components/pages/listGuides.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import api from '../../api'
-import ListItem from '../ListDisplay';
-import Search from '../search'
-
-
-
-class List extends React.Component {
-
-    
-    render() {
-        return(
-            <Search showResult />
-        )
-}
-}
-export default List;
