+++ client/src/components/search.jsx
@@ -1,9 +1,6 @@
 import React, { Component, useContext, useEffect, useState } from "react";
-
-import SearchResults from "./newSearchResult";
 import { searchGuides } from "../api/index";
 import { AppContext } from "../context/appContext";
-
 // import { ListItemAvatar } from '@material-ui/core';
 
 const Search = (props) => {
@@ -32,9 +29,9 @@ const Search = (props) => {
       language.trim().toLowerCase(),
       city.trim().toLowerCase(),
       cost.trim().toLowerCase()
-     
+
     );
-    console.log( country.trim().toLowerCase())
+    console.log(country.trim().toLowerCase())
     const guides = response.data.data;
 
     appContext.setState({ filterItems: guides });
@@ -43,22 +40,18 @@ const Search = (props) => {
     }
   };
 
-  const sortByCost = () => {
-    appContext.state.filterItems.sort((a, b) => a.cost.localeCompare(b.cost));
-    // appContext.state.filterItems.sort((a, b) => Number(a) < Number(b) ? -1 : Number(a) > Number(b) ? 1 : 0);
-    appContext.setState({ filterItems: appContext.state.filterItems });
-  };
+
 
   const { filterItems } = state;
   return (
     <div className="searchContainer">
       <section className="search-sec">
         <div className="container">
-          <form action="#" method="post" novalidate="novalidate">
+          <form action="#" method="post" noValidate="novalidate">
             <div className="row">
               <div className="col-lg-12">
                 <div className="row">
-                  <div className="col-lg-3 col-md-3 col-sm-12 p-0">
+                  <div className=" col-md-2.5 p-0">
                     <input
                       type="text"
                       name="country"
@@ -67,7 +60,7 @@ const Search = (props) => {
                       placeholder="Country"
                     />
                   </div>
-                  <div className="col-lg-3 col-md-3 col-sm-12 p-0">
+                  <div className=" col-md-2.5 p-0">
                     <input
                       type="text"
                       name="city"
@@ -76,7 +69,7 @@ const Search = (props) => {
                       placeholder="City"
                     />
                   </div>
-                  <div className="col-lg-3 col-md-3 col-sm-12 p-0">
+                  <div className=" col-md-2.5 p-0">
                     <input
                       type="text"
                       name="cost"
@@ -85,24 +78,25 @@ const Search = (props) => {
                       placeholder="Cost"
                     />
                   </div>
-                  <div className="col-lg-3 col-md-3 col-sm-12 p-0">
+                  <div className=" col-md-2.5 p-0">
                     <select
                       name="language"
                       onChange={changeHandler}
                       className="form-control search-slt"
                       id="exampleFormControlSelect1"
                     >
-                      <option>Select Language</option>
+                      <option>Language</option>
                       <option>spanish</option>
                       <option>hebrew</option>
                       <option>english</option>
                       <option>french</option>
                     </select>
                   </div>
-                  <div className="col-lg-3 col-md-3 col-sm-12 p-0">
+                  <div  className=" col-md-1 p-0">
                     <button
                       type="button"
                       onClick={getSearchResult}
+                     
                       className="btn btn-danger wrn-btn"
                     >
                       Search
@@ -114,47 +108,6 @@ const Search = (props) => {
           </form>
         </div>
       </section>
-      <div className="result-sort">
-        {appContext.state.filterItems.length != 0 ? (
-          <hgroup class="mb20">
-            <h1>Search Results</h1>
-            <h2 class="lead">
-              <strong class="text-danger">
-                {appContext.state.filterItems.length}
-              </strong>{" "}
-              results were found{" "}
-            </h2>
-          </hgroup>
-        ) : (
-          ""
-        )}
-        {appContext.state.filterItems.length != 0 ? (
-          <div>
-            <button type="button" onClick={sortByCost} className="sortBtn">
-              Price lowest first
-            </button>
-          </div>
-        ) : (
-          ""
-        )}
-      </div>
-
-      {props.showResult &&
-        appContext.state.filterItems &&
-        appContext.state.filterItems.map((item, index) => (
-          <div className="results" key={index}>
-            <SearchResults
-              pic={item.pic}
-              first_name={item.first_name}
-              last_name={item.last_name}
-              cost={item.cost}
-              country={item.country}
-              city={item.city}
-
-              // to="/"
-            />
-          </div>
-        ))}
     </div>
   );
 };
